django-server/pyproject.toml
[tool.poetry]
name = "oyucode-server"
version = "0.1.0"
description = ""
authors = ["ByamB4 <byamba4life@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
requests = "^2.25.1"
django-rest-framework = "^0.1.0"
dj-rest-auth = "^2.1.4"
django-allauth = "^0.44.0"
djangorestframework-simplejwt = "^4.6.0"
PyJWT = "1.7.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
